Add configurable instance label
Add an instance_label field to json and netstring packets, which
defaults to "hostname:port" but can be configured differently, so that
the receiving side can distinguish the origin of the log entry.  (Syslog
already reports the host name.  A place to report the port might to
allow configuring the APP-NAME field.)

Rename log_hostname variable because it clashes with the one defined in
PostgreSQL.